cambios graba agenda sips && separo operations
