+++ modules/turnos/controller/operationsCacheController/operationsAgenda.ts
@@ -10,6 +10,8 @@ import * as constantes from '../../../legacy/schemas/constantes';
 import * as logger from './../../../../utils/loggerAgendaSipsCache';
 import * as agendaSchema from '../../schemas/agenda';
 import * as turnoCtrl from './../turnoCacheController';
+import * as turnoOps from './operationsTurno';
+import * as pacienteOps from './operationsPaciente';
 
 
 // Sección de operaciones sobre MONGODB
@@ -68,7 +70,7 @@ export function checkCodificacion(agenda, pool) {
             turno = agenda.bloques[x].turnos;
 
             for (let z = 0; z < turno.length; z++) {
-                let idTurno = await existeTurnoSips(turno[z]);
+                let idTurno = await turnoOps.existeTurnoSips(turno[z]);
                 let cloneTurno: any = [];
 
                 if (idTurno) {
@@ -193,45 +195,7 @@ async function codificacionCie10(idConsulta: any, turno: any) {
 
 // Sección de operaciones sobre SIPS
 
-/**
- * @description obtiene ID de O.Social buscando coincidencias 'DNI/Cod O.S' en la tabla de PUCO
- * pudiendo devolver 0..n códigos de obra social. Según los códigos obtenidos, se retornará un único id
- * según siguiente criterio:
- *     - Si se obtiene +2 resultados, se optará por el de máxima prioridad, siendo que:
- *     - ISSN: Mínima prioridad
- *     - PAMI: Prioridad media
- *     - Cualquier otro financiador: Prioridad máxima
- *     - Si obtiene 1 resultado, es el elegido
- *     - Si se obtiene 0 resultados, se retorna el id de PLAN SUMAR por defecto, cuyo valor está en constante.
- * @param {any} documentoPaciente
- * @returns
- */
-async function getIdObraSocialSips(documentoPaciente) {
-    return new Promise(async function (resolve, reject) {
-        let transaction;
-        const idSumar = 499;
-        let query = 'SELECT TOP(1) sips_os.idObraSocial as idOS ' +
-            'FROM [Padron].[dbo].[Pd_PUCO] puco ' +
-            'JOIN [SIPS].[dbo].[Sys_ObraSocial] sips_os ON puco.CodigoOS = sips_os.cod_PUCO ' +
-            'WHERE puco.DNI =  ' + documentoPaciente +
-            'ORDER BY  ( ' +
-            'SELECT p =  ' +
-            'CASE prio.prioridad  ' +
-            'WHEN NULL THEN 1 ' +
-            'ELSE prio.prioridad ' +
-            'END ' +
-            'FROM [SIPS].[dbo].[Sys_ObraSocial_Prioridad] as prio ' +
-            'WHERE prio.idObraSocial = sips_os.idObraSocial ' +
-            ') ASC';
 
-        try {
-            let result = await new sql.Request(transaction).query(query);
-            resolve(result.length > 0 ? result[0].idOS : idSumar);
-        } catch (err) {
-            reject(err);
-        }
-    });
-}
 
 /**
  * Verifica que exista el consultorio en sips || crea el consultorio en sips
@@ -392,10 +356,10 @@ export async function guardarCacheASips(agendasMongo, index, pool) {
             datosSips.idEfector = datosArr[0][0].idEfector;
             datosSips.idProfesional = datosArr[1][0].idProfesional;
             let idAgenda = processAgenda(agenda, datosSips, pool);
-            await processTurnos(agenda, idAgenda, datosSips.idEfector);
+            await turnoOps.processTurnos(agenda, idAgenda, datosSips.idEfector);
             await checkEstadoAgenda(agenda, idAgenda);
-            await checkEstadoTurno(agenda, idAgenda);
-            await checkAsistenciaTurno(agenda);
+            await turnoOps.checkEstadoTurno(agenda, idAgenda);
+            await turnoOps.checkAsistenciaTurno(agenda);
 
             transaccion.commit(async err2 => {
                 if (err2) {
@@ -465,136 +429,6 @@ function getEstadoAgenda(idAgenda: any) {
                 reject(err);
             }
         })();
-    });
-}
-
-function checkAsistenciaTurno(agenda: any) {
-    let turnos;
-    return new Promise(async function (resolve, reject) {
-        try {
-            for (let x = 0; x < agenda.bloques.length; x++) {
-                turnos = agenda.bloques[x].turnos;
-
-                for (let i = 0; i < turnos.length; i++) {
-                    if (turnos[i].asistencia === 'asistio') {
-
-                        let idTurno: any = await getEstadoTurnoSips(turnos[i]._id);
-                        let fechaAsistencia = moment(turnos[i].updatedAt).format('YYYYMMDD');
-                        let query = 'INSERT INTO dbo.CON_TurnoAsistencia ( idTurno , idUsuario , fechaAsistencia ) VALUES  ( ' +
-                            idTurno.idTurno + ' , ' + constantes.idUsuarioSips + ' , \'' + fechaAsistencia + '\' )';
-
-                        await executeQuery(query);
-
-                        // resolve();
-                    } else {
-                        // resolve();
-                    }
-                }
-            }
-            resolve();
-        } catch (ex) {
-            reject(ex);
-        }
-    });
-}
-
-function checkEstadoTurno(agenda: any, idAgendaSips) {
-    let turnos;
-    return new Promise(async function (resolve, reject) {
-        try {
-            for (let x = 0; x < agenda.bloques.length; x++) {
-                turnos = agenda.bloques[x].turnos;
-
-                for (let i = 0; i < turnos.length; i++) {
-                    if ((turnos[i].estado !== 'disponible') || (turnos[i].updatedAt)) {
-                        await actualizarEstadoTurnoSips(idAgendaSips, turnos[i]);
-                        // resolve();
-                    } else {
-                        // resolve();
-                    }
-                }
-            }
-            resolve();
-
-        } catch (ex) {
-            reject(ex);
-        }
-    });
-}
-
-function processTurnos(agendas: any, idAgendaCreada: any, idEfector: any) {
-    let turnos;
-    return new Promise(async function (resolve, reject) {
-        try {
-            for (let x = 0; x < agendas.bloques.length; x++) {
-                turnos = agendas.bloques[x].turnos;
-
-                for (let i = 0; i < turnos.length; i++) {
-
-                    if (turnos[i].estado === 'asignado') {
-                        let idTurno = await existeTurnoSips(turnos[i]);
-
-                        if (!idTurno) {
-                            await grabaTurnoSips(turnos[i], idAgendaCreada, idEfector);
-                            // resolve();
-                        } else {
-                            // resolve();
-                        }
-                        // resolve();
-                    } else {
-                        // resolve();
-                    }
-                }
-            }
-            resolve();
-        } catch (ex) {
-            reject(ex);
-        }
-    });
-}
-async function grabaTurnoSips(turno, idAgendaSips, idEfector) {
-    return new Promise(async function (resolve, reject) {
-        try {
-            // TODO: El paciente pudiera no estar validado, en ese caso no se encontrara en la
-            // colección de paciente de MPI, en ese caso buscar en la coleccion de pacientes de Andes
-            let pacienteEncontrado = await pacientes.buscarPaciente(turno.paciente.id);
-            let paciente = pacienteEncontrado.paciente;
-            // if(!pacienteEncontrado) {
-            //  pacienteEncontrado = buscar en andes.......
-            // }
-
-            let idObraSocial = await getIdObraSocialSips(paciente.documento);
-            let pacienteId = await getPacienteMPI(paciente, idEfector);
-
-            let fechaTurno = moment(turno.horaInicio).format('YYYYMMDD');
-            let horaTurno = moment(turno.horaInicio).utcOffset('-03:00').format('HH:mm');
-
-            let query = 'INSERT INTO dbo.CON_Turno ( idAgenda , idTurnoEstado , idUsuario ,  idPaciente ,  fecha , hora , sobreturno , idTipoTurno , idObraSocial , idTurnoAcompaniante, objectId ) VALUES  ( ' +
-                idAgendaSips + ' , 1 , ' + constantes.idUsuarioSips + ' ,' + pacienteId + ', \'' + fechaTurno + '\' ,\'' + horaTurno + '\' , 0 , 0 ,' + idObraSocial + ' , 0, \'' + turno._id + '\')';
-
-            let turnoGrabado = await executeQuery(query);
-            resolve(turnoGrabado);
-        } catch (ex) {
-            reject(ex);
-        }
-    });
-}
-
-export function existeTurnoSips(turno: any) {
-    return new Promise(function (resolve, reject) {
-        let transaction;
-        return new sql.Request(transaction)
-            .input('idTurnoMongo', sql.VarChar(50), turno._id)
-            .query('SELECT idTurno FROM dbo.CON_Turno WHERE objectId = @idTurnoMongo GROUP BY idTurno')
-            .then(result => {
-                if (result.length > 0) {
-                    resolve(result[0].idTurno);
-                } else {
-                    resolve(false);
-                }
-            }).catch(err => {
-                reject(err);
-            });
     });
 }
 
@@ -685,7 +519,6 @@ async function grabaAgendaSips(agendaSips: any, datosSips: any, pool) {
 
     let dniProfesional = agendaSips.profesionales[0].documento;
     try {
-        let listaIdProfesionales = await Promise.all(getProfesionalesSips(agendaSips.profesionales, pool));
 
         if (agendaSips.profesionales.length > 1) {
             multiprofesional = 1;
@@ -696,37 +529,32 @@ async function grabaAgendaSips(agendaSips: any, datosSips: any, pool) {
         let idEfector = datosSips.idEfector;
         let idProfesional = datosSips.idProfesional;
         let idEspecialidad = (agendaSips.tipoPrestaciones[0].term.includes('odonto')) ? 34 : 14;  /*IdEspecialidad 34 = odontologia en SIPS*/
-
         let idServicio = 177;
         let idTipoPrestacion = 0;
         let idConsultorio = await existeConsultorio(agendaSips, idEfector);
-
+        // ---> Grabamos la agenda en SIPS
         let query = 'insert into Con_Agenda (idAgendaEstado, idEfector, idServicio, idProfesional, idTipoPrestacion, idEspecialidad, idConsultorio, fecha, duracion, horaInicio, horaFin, maximoSobreTurnos, porcentajeTurnosDia, porcentajeTurnosAnticipados, citarPorBloques, cantidadInterconsulta, turnosDisponibles, idMotivoInactivacion, multiprofesional, objectId) ' +
             'values (' + estado + ', ' + idEfector + ', ' + idServicio + ', ' + idProfesional + ', ' + idTipoPrestacion + ', ' + idEspecialidad + ', ' + idConsultorio + ', \'' + fecha + '\', ' + duracionTurno + ', \'' + horaInicio + '\', \'' + horaFin + '\', ' + maximoSobreTurnos + ', ' + porcentajeTurnosDia + ', ' + porcentajeTurnosAnticipados + ', ' + citarPorBloques + ' , ' + cantidadInterconsulta + ', ' + turnosDisponibles + ', ' + idMotivoInactivacion + ', ' + multiprofesional + ', \'' + objectId + '\')';
 
-        executeQuery(query).then(function (idAgendaCreada) {
-            let query2;
+        let idAgendaCreada = await executeQuery(query);
 
-            if (listaIdProfesionales.length > 0) {
-                listaIdProfesionales.forEach(async function (listaIdProf) {
-
-                    query2 = 'INSERT INTO dbo.CON_AgendaProfesional ( idAgenda, idProfesional, baja, CreatedBy , ' +
-                        ' CreatedOn, ModifiedBy, ModifiedOn, idEspecialidad ) VALUES  ( ' +
-                        idAgendaCreada + ',' +
-                        listaIdProf[0].idProfesional + ',' +
-                        0 + ',' +
-                        constantes.idUsuarioSips + ',' +
-                        '\'' + moment().format('YYYYMMDD HH:mm:ss') + '\', ' +
-                        '\'' + moment().format('YYYYMMDD HH:mm:ss') + '\', ' +
-                        '\'' + moment().format('YYYYMMDD HH:mm:ss') + '\', ' +
-                        idEspecialidad + ' ) ';
-
-                    await executeQuery(query2);
-                });
-            }
-
-            return (idAgendaCreada);
-        });
+        // ---> Obtenemos los id's de profesionales(SIPS) para la agenda actual y luego insertamos las "agendasProfesional" correspondientes en SIPS
+        let listaIdProfesionales = await Promise.all(getProfesionales(agendaSips.profesionales, pool));
+        if (listaIdProfesionales.length > 0) {
+            let promiseArray = [];
+            listaIdProfesionales.forEach(listaIdProf => {
+                let query2 = 'select SCOPE_IDENTITY() as id INSERT INTO dbo.CON_AgendaProfesional ( idAgenda, idProfesional, baja, CreatedBy , ' +
+                    ' CreatedOn, ModifiedBy, ModifiedOn, idEspecialidad ) VALUES  ( ' + idAgendaCreada + ',' +
+                    listaIdProf[0].idProfesional + ',' + 0 + ',' + constantes.idUsuarioSips + ',' +
+                    '\'' + moment().format('YYYYMMDD HH:mm:ss') + '\', ' +
+                    '\'' + moment().format('YYYYMMDD HH:mm:ss') + '\', ' +
+                    '\'' + moment().format('YYYYMMDD HH:mm:ss') + '\', ' +
+                    idEspecialidad + ' ) ';
+                promiseArray.push(new sql.request().query(query2));
+            });
+            await Promise.all(promiseArray);
+        }
+        return (idAgendaCreada);
     } catch (err) {
         console.log('-----------------> ERROR en existeAgendaSips ', err);
         return err;
@@ -744,14 +572,15 @@ function getEstadoAgendaSips(estadoCitas) {
     }
     return (estado);
 }
+
 /**
-* Busca profesionales SIPS para cada profesional en la agenda de Mongo, devuelve un array de promises.
+* Busca los id's en SIPS de los profesionales de la agenda y devuelve un array de promises.
  *
  * @param {any} profesionalesMongo
  * @param {any} pool
  * @returns
  */
-function getProfesionalesSips(profesionalesMongo, pool) {
+function getProfesionales(profesionalesMongo, pool) {
     let profesionalesSipsPromise = [];
     profesionalesMongo.map(async profMongo => profesionalesSipsPromise.push(arrayIdProfesionales(profMongo, pool)));
     return profesionalesSipsPromise;
@@ -762,6 +591,7 @@ function arrayIdProfesionales(profMongo, pool) {
         .input('dniProfesional', sql.Int, profMongo.documento)
         .query('SELECT idProfesional FROM dbo.Sys_Profesional WHERE numeroDocumento = @dniProfesional AND activo = 1');
 }
+
 /**
  * @param agenda @description Busca el paciente correspondiente a un turno y agenda pasados por parámetro
  * @param idTurno
@@ -787,43 +617,6 @@ function getPacienteAgenda(agenda, idTurno) {
             }
             resolve();
         });
-    });
-}
-async function actualizarEstadoTurnoSips(idAgendaSips, turno) {
-    return new Promise(async function (resolve, reject) {
-        try {
-            let estadoTurnoSips: any = await getEstadoTurnoSips(turno._id);
-            let estadoTurnoMongo = await getEstadoTurnosCitasSips(turno.estado, turno.updatedAt);
-
-            if (estadoTurnoSips.idTurnoEstado !== estadoTurnoMongo) {
-                let objectIdTurno;
-
-                if (turno._id) {
-                    objectIdTurno = ' and objectId = \'' + turno._id + '\'';
-                    resolve();
-                } else {
-                    resolve();
-                }
-
-                /*TODO: hacer enum con los estados */
-                var horaInicio = moment(turno.horaInicio).utcOffset('-03:00').format('HH:mm');
-
-                if ((estadoTurnoMongo === constantes.EstadoTurnosSips.suspendido || turno.estado === 'turnoDoble') && !await existeTurnoBloqueoSips(idAgendaSips, horaInicio)) {
-                    await grabarTurnoBloqueo(idAgendaSips, turno);
-                    resolve();
-                } else {
-                    resolve();
-                }
-
-                let query = 'UPDATE dbo.CON_Turno SET idTurnoEstado = ' + estadoTurnoMongo + ' WHERE idAgenda = ' + idAgendaSips + objectIdTurno;
-                await executeQuery(query);
-                resolve();
-            } else {
-                resolve();
-            }
-        } catch (ex) {
-            reject(ex);
-        }
     });
 }
 
@@ -939,29 +732,7 @@ function getEstadoTurnosCitasSips(estadoTurnoCitas, updated) {
     });
 }
 
-/* Devuelve el estado del turno en Con_Turno de SIPS */
-function getEstadoTurnoSips(objectId: any) {
-    return new Promise((resolve: any, reject: any) => {
-        let transaction;
-        (async function () {
-            try {
-                let query = 'SELECT idAgenda, idTurno, idTurnoEstado FROM dbo.CON_Turno WHERE objectId = @objectId';
-                let result = await new sql.Request(transaction)
-                    .input('objectId', sql.VarChar(50), objectId)
-                    .query(query);
 
-                if (typeof result[0] !== 'undefined') {
-                    resolve(result[0]);
-                } else {
-                    let idTurnoEstado = 0;
-                    resolve(idTurnoEstado);
-                }
-            } catch (err) {
-                reject(err);
-            }
-        })();
-    });
-}
 
 function executeQuery(query: any) {
     query += ' select SCOPE_IDENTITY() as id';
@@ -976,225 +747,7 @@ function executeQuery(query: any) {
     });
 }
 
-function existePacienteSips(pacienteSips) {
-    let idPacienteSips;
-    return new Promise((resolve: any, reject: any) => {
-        let transaction;
-        (async function () {
-            try {
-                let query = 'SELECT idPaciente FROM dbo.Sys_Paciente WHERE objectId = @objectId';
-                let result = await new sql.Request(transaction)
-                    .input('objectId', sql.VarChar(50), pacienteSips.objectId)
-                    .query(query);
 
-                if (typeof result[0] !== 'undefined') {
-                    idPacienteSips = result[0].idPaciente;
-                    resolve(idPacienteSips);
-                } else {
-                    idPacienteSips = 0;
-                    resolve(idPacienteSips);
-                }
 
-            } catch (err) {
-                reject(err);
-            }
-        })();
-    });
-}
 
-async function insertaPacienteSips(pacienteSips: any) {
-    return new Promise(async function (resolve, reject) {
-        try {
-            let idPacienteGrabadoSips;
-            let idPaciente = await existePacienteSips(pacienteSips);
-
-            if (idPaciente === 0) {
-
-                let query = 'INSERT INTO dbo.Sys_Paciente ' +
-                    ' ( idEfector ,' +
-                    ' apellido , ' +
-                    ' nombre, ' +
-                    ' numeroDocumento, ' +
-                    ' idSexo, ' +
-                    ' fechaNacimiento, ' +
-                    ' idEstado, ' +
-                    ' idMotivoNI, ' +
-                    ' idPais, ' +
-                    ' idProvincia, ' +
-                    ' idNivelInstruccion, ' +
-                    ' idSituacionLaboral, ' +
-                    ' idProfesion, ' +
-                    ' idOcupacion, ' +
-                    ' calle, ' +
-                    ' numero, ' +
-                    ' piso, ' +
-                    ' departamento, ' +
-                    ' manzana, ' +
-                    ' idBarrio, ' +
-                    ' idLocalidad, ' +
-                    ' idDepartamento, ' +
-                    ' idProvinciaDomicilio, ' +
-                    ' referencia, ' +
-                    ' informacionContacto, ' +
-                    ' cronico, ' +
-                    ' idObraSocial, ' +
-                    ' idUsuario, ' +
-                    ' fechaAlta, ' +
-                    ' fechaDefuncion, ' +
-                    ' fechaUltimaActualizacion, ' +
-                    ' idEstadoCivil, ' +
-                    ' idEtnia, ' +
-                    ' idPoblacion, ' +
-                    ' idIdioma, ' +
-                    ' otroBarrio, ' +
-                    ' camino, ' +
-                    ' campo, ' +
-                    ' esUrbano, ' +
-                    ' lote, ' +
-                    ' parcela, ' +
-                    ' edificio, ' +
-                    ' activo, ' +
-                    ' fechaAltaObraSocial, ' +
-                    ' numeroAfiliado, ' +
-                    ' numeroExtranjero, ' +
-                    ' telefonoFijo, ' +
-                    ' telefonoCelular, ' +
-                    ' email, ' +
-                    ' latitud, ' +
-                    ' longitud, ' +
-                    ' objectId ) ' +
-                    ' VALUES( ' +
-                    pacienteSips.idEfector + ', ' +
-                    '\'' + pacienteSips.apellido + '\',' +
-                    '\'' + pacienteSips.nombre + '\', ' +
-                    pacienteSips.numeroDocumento + ', ' +
-                    pacienteSips.idSexo + ', ' +
-                    '\'' + pacienteSips.fechaNacimiento + '\',' +
-                    pacienteSips.idEstado + ', ' +
-                    pacienteSips.idMotivoNI + ', ' +
-                    pacienteSips.idPais + ', ' +
-                    pacienteSips.idProvincia + ', ' +
-                    pacienteSips.idNivelInstruccion + ', ' +
-                    pacienteSips.idSituacionLaboral + ', ' +
-                    pacienteSips.idProfesion + ', ' +
-                    pacienteSips.idOcupacion + ', ' +
-                    '\'' + pacienteSips.calle + '\', ' +
-                    pacienteSips.numero + ', ' +
-                    '\'' + pacienteSips.piso + '\', ' +
-                    '\'' + pacienteSips.departamento + '\', ' +
-                    '\'' + pacienteSips.manzana + '\', ' +
-                    pacienteSips.idBarrio + ', ' +
-                    pacienteSips.idLocalidad + ', ' +
-                    pacienteSips.idDepartamento + ', ' +
-                    pacienteSips.idProvinciaDomicilio + ', ' +
-                    '\'' + pacienteSips.referencia + '\', ' +
-                    '\'' + pacienteSips.informacionContacto + '\', ' +
-                    pacienteSips.cronico + ', ' +
-                    pacienteSips.idObraSocial + ', ' +
-                    pacienteSips.idUsuario + ', ' +
-                    '\'' + pacienteSips.fechaAlta + '\', ' +
-                    '\'' + pacienteSips.fechaDefuncion + '\', ' +
-                    '\'' + pacienteSips.fechaUltimaActualizacion + '\', ' +
-                    pacienteSips.idEstadoCivil + ', ' +
-                    pacienteSips.idEtnia + ', ' +
-                    pacienteSips.idPoblacion + ', ' +
-                    pacienteSips.idIdioma + ', ' +
-                    '\'' + pacienteSips.otroBarrio + '\', ' +
-                    '\'' + pacienteSips.camino + '\', ' +
-                    '\'' + pacienteSips.campo + '\', ' +
-                    pacienteSips.esUrbano + ', ' +
-                    '\'' + pacienteSips.lote + '\', ' +
-                    '\'' + pacienteSips.parcela + '\', ' +
-                    '\'' + pacienteSips.edificio + '\', ' +
-                    pacienteSips.activo + ', ' +
-                    '\'' + pacienteSips.fechaAltaObraSocial + '\', ' +
-                    '\'' + pacienteSips.numeroAfiliado + '\', ' +
-                    '\'' + pacienteSips.numeroExtranjero + '\', ' +
-                    '\'' + pacienteSips.telefonoFijo + '\', ' +
-                    '\'' + pacienteSips.telefonoCelular + '\', ' +
-                    '\'' + pacienteSips.email + '\', ' +
-                    '\'' + pacienteSips.latitud + '\', ' +
-                    '\'' + pacienteSips.longitud + '\', ' +
-                    '\'' + pacienteSips.objectId + '\' ' +
-                    ') ';
-
-                idPacienteGrabadoSips = await executeQuery(query);
-            } else {
-                idPacienteGrabadoSips = idPaciente;
-            }
-
-            resolve(idPacienteGrabadoSips);
-        } catch (ex) {
-            reject(ex);
-        }
-    });
-}
 // #region GetPacienteMPI
-/** Este método se llama desde grabaTurnoSips */
-async function getPacienteMPI(paciente: any, idEfectorSips: any) {
-    return new Promise(async function (resolve, reject) {
-        try {
-            let pacienteSips = {
-                idEfector: idEfectorSips,
-                nombre: paciente.nombre,
-                apellido: paciente.apellido,
-                numeroDocumento: paciente.documento,
-                idSexo: (paciente.sexo === 'masculino' ? 3 : paciente.sexo === 'femenino' ? 2 : 1),
-                fechaNacimiento: moment(paciente.fechaNacimiento).format('YYYYMMDD'),
-                idEstado: 3,
-                /* Estado Validado en SIPS*/
-                idMotivoNI: 0,
-                idPais: 54,
-                idProvincia: 139,
-                idNivelInstruccion: 0,
-                idSituacionLaboral: 0,
-                idProfesion: 0,
-                idOcupacion: 0,
-                calle: '',
-                numero: 0,
-                piso: '',
-                departamento: '',
-                manzana: '',
-                idBarrio: -1,
-                idLocalidad: 52,
-                idDepartamento: 557,
-                idProvinciaDomicilio: 139,
-                referencia: '',
-                informacionContacto: '',
-                cronico: 0,
-                idObraSocial: 499,
-                idUsuario: constantes.idUsuarioSips,
-                fechaAlta: moment().format('YYYYMMDD HH:mm:ss'),
-                fechaDefuncion: '19000101',
-                fechaUltimaActualizacion: moment().format('YYYYMMDD HH:mm:ss'),
-                idEstadoCivil: 0,
-                idEtnia: 0,
-                idPoblacion: 0,
-                idIdioma: 0,
-                otroBarrio: '',
-                camino: '',
-                campo: '',
-                esUrbano: 1,
-                lote: '',
-                parcela: '',
-                edificio: '',
-                activo: 1,
-                fechaAltaObraSocial: '19000101',
-                numeroAfiliado: null,
-                numeroExtranjero: '',
-                telefonoFijo: 0,
-                telefonoCelular: 0,
-                email: '',
-                latitud: 0,
-                longitud: 0,
-                objectId: paciente._id
-            };
-
-            let idPacienteGrabadoSips = await insertaPacienteSips(pacienteSips);
-
-            resolve(idPacienteGrabadoSips);
-        } catch (ex) {
-            reject(ex);
-        }
-    });
-}
